Fixes bug where WE v5 is used with WG v6.
WorldGuardInterface is now instantiated based on the WorldEdit version
that the server is running (if any). Instead of the WorldGuard version.

The WorldGuardInterface uses WorldEdit behind the scenes to perform some
of its responsibilities for saving regions and resetting regions.

methodName() == methodIntention()
--------------------------------------
saveSchematic() == saveArenaRegion()
pasteSchematic() == resetArenaRegion()

These methods are located under the WorldGuard Utilities & Interface
because of their intention for saving and resetting worldguard regions.