Adds anonymous field tester.
The job of the Tester is to make sure that all WorldGuard fields are not
null. This means that WG is fully initialized.

The problem comes from WG v6.0.0-beta where can exception occurs during
onEnable() about line 200 when run on Spigot/Craftbukkit 1.8 (because of
the Guava update and the Guava package has moved to a new location inside
the JAR). This means that flagStateManager on line 219 does NOT get
initialized, and no WG events are registered with Bukkit either.